Adicionado novas issues no controller de funcionarios

+++ src/funcionario/funcionario.service.ts
@@ -1,15 +1,20 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Funcionario } from '../models/FuncionarioEntity';
-import { CreateFuncionarioDto } from './dto/create-funcionario.dto';
+import {
+  CreateFuncionarioDto,
+  SaveFuncionarioDto,
+} from './dto/create-funcionario.dto';
 import { UpdateFuncionarioDto } from './dto/update-funcionario.dto';
+import { CompanhiaService } from '../companhia/companhia.service';
 
 @Injectable()
 export class FuncionarioService {
   constructor(
     @InjectRepository(Funcionario)
     private funcionariosRepository: Repository<Funcionario>,
+    private companhiaService: CompanhiaService,
   ) {}
 
   findByCompanhia(idCompanhia: number) {
@@ -19,8 +24,38 @@ export class FuncionarioService {
     });
   }
 
-  create(createFuncionarioDto: CreateFuncionarioDto) {
-    return this.funcionariosRepository.save(createFuncionarioDto);
+  async create(createFuncionarioDto: CreateFuncionarioDto) {
+    if (!createFuncionarioDto.companhia) {
+      throw new HttpException(
+        {
+          status: HttpStatus.FORBIDDEN,
+          error: 'Campo >companhia< deve ser definido!',
+        },
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
+    const companiaExiste = await this.companhiaService.findOne(
+      createFuncionarioDto.companhia.toString(),
+    );
+
+    if (!companiaExiste) {
+      throw new HttpException(
+        {
+          status: HttpStatus.FORBIDDEN,
+          error: 'Companhia id não encontrada!',
+        },
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
+    const saveFuncionario: SaveFuncionarioDto = {
+      nome: createFuncionarioDto.nome,
+      email: createFuncionarioDto.email,
+      hashSenha: createFuncionarioDto.hashSenha,
+      companhia: companiaExiste,
+    }
+    return this.funcionariosRepository.save(saveFuncionario);
   }
 
   findAll() {
@@ -42,6 +77,19 @@ export class FuncionarioService {
   }
 
   async remove(id: number) {
-    return this.funcionariosRepository.delete(id);
+    const funcionario = await this.funcionariosRepository.findOne(id);
+    if (!funcionario) {
+      throw new HttpException(
+        {
+          status: HttpStatus.FORBIDDEN,
+          error: 'Funcionário não encontrado.',
+        },
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
+    await this.funcionariosRepository.delete(id);
+
+    return { msg: 'Funcionário deletado com sucesso!' };
   }
 }
